Moved quickslots classes under modules.
