+++ Outward.Shared/ModifAmorphic.Outward.Shared/Modules/Character/SkillSchools/Patches/SkillTreeHolderPatches.cs
@@ -3,14 +3,14 @@ using ModifAmorphic.Outward.Events;
 using ModifAmorphic.Outward.Logging;
 using System;
 
-namespace ModifAmorphic.Outward.SkillSchools.Patches
+namespace ModifAmorphic.Outward.Modules.Character
 {
-    [HarmonyPatch(typeof(CharacterManager))]
-    internal static class CharacterManagerPatches
+    [HarmonyPatch(typeof(SkillTreeHolder))]
+    internal static class SkillTreeHolderPatches
     {
         private static Func<IModifLogger> _getLogger;
         private static IModifLogger Logger => _getLogger?.Invoke() ?? new NullLogger();
-        public static event Action<CharacterManager> AwakeAfter;
+        public static event Action<SkillTreeHolder> AwakeAfter;
 
 
         [EventSubscription]
@@ -21,7 +21,7 @@ namespace ModifAmorphic.Outward.SkillSchools.Patches
 
         [HarmonyPatch("Awake", MethodType.Normal)]
         [HarmonyPostfix]
-        private static void AwakePostfix(CharacterManager __instance)
+        private static void AwakePostfix(SkillTreeHolder __instance)
         {
             AwakeAfter?.Invoke(__instance);
         }
