+++ Outward.Shared/ModifAmorphic.Outward.Shared/Modules/Character/SkillSchools/SchoolService.cs
@@ -1,11 +1,10 @@
 ﻿using ModifAmorphic.Outward.Logging;
-using ModifAmorphic.Outward.Modules.Instances;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace ModifAmorphic.Outward.SkillSchools
+namespace ModifAmorphic.Outward.Modules.Character
 {
     public class SchoolService
     {
