+++ Outward.Shared/ModifAmorphic.Outward.Shared/Modules/QuickSlots/KeyBindings/Patches/LocalizationManagerPatches.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ModifAmorphic.Outward.KeyBindings.Patches
+namespace ModifAmorphic.Outward.Modules.QuickSlots.KeyBindings
 {
     [HarmonyPatch(typeof(LocalizationManager), "Awake")]
     internal static class LocalizationManagerPatches
